ecommerce-template/requirements.txt
Django==1.11
Pillow==5.2.0
boto3==1.9.9
botocore==1.12.9
dj-database-url==0.5.0
django-forms-bootstrap==3.1.0
django-storages==1.7.1
docutils==0.14
gunicorn==19.9.0
jmespath==0.9.3
psycopg2==2.7.5
python-dateutil==2.7.3
pytz==2018.5
s3transfer==0.1.13
stripe==2.8.1
urllib3==1.23
